Add route for deleting a user account

+++ app/routers/db.py
@@ -1,4 +1,4 @@
-from .db_helpers.models import UserCreateResp, UserLoginReq, UserLoginResp, TeamGetResp, TeamAddReq, TeamAddResp, TeamRemoveReq, TeamRemoveResp, TeamUpdateReq, TeamUpdateResp, UserUpdateReq, UserUpdateResp, UserDeleteResp, GenerateLineupReq, GenerateLineupResp, SaveLineupReq, SaveLineupResp, GetLineupsResp, DeleteLineupResp, VerifyEmailReq, CheckCodeReq
+from .db_helpers.models import UserCreateResp, UserLoginReq, UserLoginResp, TeamGetResp, TeamAddReq, TeamAddResp, TeamRemoveReq, TeamRemoveResp, TeamUpdateReq, TeamUpdateResp, UserUpdateReq, UserUpdateResp, UserDeleteResp, GenerateLineupReq, GenerateLineupResp, SaveLineupReq, SaveLineupResp, GetLineupsResp, DeleteLineupResp, VerifyEmailReq, CheckCodeReq, UserDeleteReq
 from .db_helpers.utils import hash_password, check_password, create_access_token, get_current_user, serialize_league_info, serialize_lineup_info, generate_lineup_hash, deserialize_lineups, generate_verification_code, send_verification_email
 from .constants import ACCESS_TOKEN_EXPIRE_DAYS, FEATURES_SERVER_ENDPOINT, DB_CREDENTIALS
 from .data_helpers.utils import check_league
@@ -228,10 +228,17 @@ async def view_team(team_id: int, current_user: dict = Depends(get_current_user)
 # ------------------------------------ User Management -------------------------------------- #
 
 @router.post('/users/delete')
-async def delete_user(current_user: dict = Depends(get_current_user)):
+async def delete_user(user: UserDeleteReq, current_user: dict = Depends(get_current_user)):
 	user_id = current_user.get("uid")
+	password = user.password
 
 	with get_cursor() as cur:
+		cur.execute("SELECT password FROM users WHERE uid = %s LIMIT 1", (user_id,))
+		user_data = cur.fetchone()
+
+		if not user_data or not check_password(password, user_data[0]):
+			return UserDeleteResp(success=False)
+
 		cur.execute("DELETE FROM users WHERE user_id = %s", (user_id,))
 		cur.execute("DELETE FROM teams WHERE user_id = %s", (user_id,))
 		conn.commit()
